requirements.txt
aniso8601==9.0.1
certifi==2018.1.18
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.0.3
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
svgwrite==1.4.1
torch==1.8.1
torchaudio==0.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
webcolors==1.11.1
Werkzeug==1.0.1
Flask-Cors==3.0.10